+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Japanese translation @terapyon
 - Fix `ArrayWidget` to support multiselect schema `schema.List`/`schema.Set`-> `schema.Choice` hardcoded (not using vocabularies) combination @sneridagh
+- Show past dates in date time widget calendar @nzambello
 
 ### Internal